+++ weblate/settings_example.py
@@ -718,6 +718,7 @@ CRISPY_TEMPLATE_PACK = 'bootstrap3'
 #     'weblate.checks.markup.MarkdownLinkCheck',
 #     'weblate.checks.markup.MarkdownSyntaxCheck',
 #     'weblate.checks.markup.URLCheck',
+#     'weblate.checks.markup.SafeHTMLCheck',
 #     'weblate.checks.placeholders.PlaceholderCheck',
 #     'weblate.checks.placeholders.RegexCheck',
 #     'weblate.checks.source.OptionalPluralCheck',