+++ weblate/trans/models/_conf.py
@@ -66,6 +66,7 @@ class WeblateConf(AppConf):
         'weblate.trans.autofixes.chars.ReplaceTrailingDotsWithEllipsis',
         'weblate.trans.autofixes.chars.RemoveZeroSpace',
         'weblate.trans.autofixes.chars.RemoveControlChars',
+        'weblate.trans.autofixes.html.BleachHTML',
     )
 
     # Matomo